refactor: separate library and command
