+++ .gitignore
@@ -22,3 +22,4 @@ go.work
 
 # binary
 godocx
+!godocx/
